requirements.txt
fastapi==0.100.0

uvicorn==0.23.2

pymongo==4.5.0

python-dotenv==1.0.0

PyYAML==6.0.1

streamlit==1.26.0

pandas==2.2.2

numpy==1.26.4

matplotlib==3.8.2

seaborn==0.13.2

plotly==5.18.0